web/backend/requirements.txt
# ----- Web Framework -----
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# ----- Data Validation -----
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ----- Environment -----
python-dotenv>=1.0.0

# ----- AI & RAG Pipeline -----
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
google-generativeai>=0.3.0

# ----- Vector Database -----
faiss-cpu>=1.7.4

# ----- Database -----
motor==3.3.2
pymongo==4.6.1

# ----- File Upload -----
python-multipart>=0.0.6

# ----- Utilities -----
httpx>=0.25.0
